+++ src/bin/bat/app.rs
@@ -121,7 +121,7 @@ impl App {
             _ => unreachable!("other values for --paging are not allowed"),
         };
 
-        let mut syntax_mapping = SyntaxMapping::builtin();
+        let mut syntax_mapping = SyntaxMapping::new();
 
         if let Some(values) = self.matches.get_many::<String>("ignored-suffix") {
             for suffix in values {