+++ CHANGELOG.md
@@ -16,6 +16,8 @@
 - Update Arch Linux package URL in README files #2779 (@brunobell)
 - Update and improve `zsh` completion, see #2772 (@okapia)
 - More extensible syntax mapping mechanism #2755 (@cyqsimon)
+  - [BREAKING] Precedence order of user-defined syntax mappings (using `-m/--map-syntax`) has been reversed
+    - `-m '*.foo:alpha' -m '*.foo:bravo'` now maps `bar.foo` to `alpha`
 
 ## Syntaxes
 
@@ -23,6 +25,10 @@
 
 ## `bat` as a library
 
+- Changes to `syntax_mapping::SyntaxMapping` #2755 (@cyqsimon)
+  - `SyntaxMapping::get_syntax_for` is now correctly public
+  - [BREAKING] `SyntaxMapping::{empty,builtin}` are removed; use `SyntaxMapping::new` instead
+  - [BREAKING] `SyntaxMapping::mappings` is replaced by `SyntaxMapping::{builtin,custom,all}_mappings`
 
 # v0.24.0